避免不必要的cast操作

+++ vllm_ascend/ops/fused_moe/experts_selector.py
@@ -242,7 +242,7 @@ def _select_experts_with_fusion_ops(
             forward_context = get_forward_context()
             if input_ids is None:
                 input_ids = forward_context.input_ids
-            input_ids = input_ids.to(torch.int64)
+            input_ids = input_ids.to(torch.int32)
             # tid2eid_ones = torch.ones(tid2eid.shape[0],tid2eid.shape[1],device=router_logits.device,dtype=torch.int32)
             tid2eid_ones = tid2eid.to(torch.int32)
             if forward_context.moe_comm_type == MoECommType.ALLGATHER:
